backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
